+++ React/src/components/SelectProduct.tsx
@@ -101,8 +101,92 @@
 
 // export default SelectProduct;
 
+
+
+// import { Card, CardBody, Select, Text } from "@chakra-ui/react";
+// import React, { useState, useEffect } from "react";
+
+// interface Product {
+//   name: string;
+//   description: string;
+//   store_id: number;
+//   price: number;
+//   stock: number;
+//   id: number;
+// }
+
+// interface SelectedProduct extends Product {
+//   quantity: number;
+// }
+
+// interface ProductListProps {
+//   selectedGroup: number;
+//   selectedProducts: SelectedProduct[];
+//   onProductSelect: (product: SelectedProduct, quantity: number) => void;
+// }
+
+// const SelectProduct: React.FC<ProductListProps> = ({
+//   selectedGroup,
+//   selectedProducts,
+//   onProductSelect,
+// }) => {
+//   const [products, setProducts] = useState<Product[]>([]);
+
+//   useEffect(() => {
+//     const fetchProducts = async () => {
+//       const response = await fetch("http://127.0.0.1:8000/products");
+//       const data = await response.json();
+//       const filteredProducts = data.filter(
+//         (product: Product) => product.store_id === selectedGroup
+//       );
+//       setProducts(filteredProducts);
+//     };
+
+//     fetchProducts();
+//   }, [selectedGroup]);
+
+//   const handleQuantityChange = (product: Product, quantity: number) => {
+//     onProductSelect({ ...product, quantity }, quantity);
+//   };
+
+//   return (
+//     <div>
+//       {products.map((product) => {
+//         const selectedProduct = selectedProducts.find(
+//           (p) => p.id === product.id
+//         );
+//         const initialQuantity = selectedProduct ? selectedProduct.quantity : 0;
+
+//         return (
+//           <Card key={product.id} mt="5">
+//             <CardBody>
+//               <Text>{product.name}</Text>
+//               <Select
+//                 value={initialQuantity}
+//                 onChange={(e) =>
+//                   handleQuantityChange(product, parseInt(e.target.value, 10))
+//                 }
+//               >
+//                 {[...Array(product.stock + 1).keys()].map((i) => (
+//                   <option key={i} value={i}>
+//                     {i}
+//                   </option>
+//                 ))}
+//               </Select>
+//             </CardBody>
+//           </Card>
+//         );
+//       })}
+//     </div>
+//   );
+// };
+
+// export default SelectProduct;
+
+
+
 import { Card, CardBody, Select, Text } from "@chakra-ui/react";
-import React, { useState, useEffect } from "react";
+import React from "react";
 
 interface Product {
   name: string;
@@ -121,28 +205,15 @@ interface ProductListProps {
   selectedGroup: number;
   selectedProducts: SelectedProduct[];
   onProductSelect: (product: SelectedProduct, quantity: number) => void;
+  products: Product[];
 }
 
 const SelectProduct: React.FC<ProductListProps> = ({
   selectedGroup,
   selectedProducts,
   onProductSelect,
+  products
 }) => {
-  const [products, setProducts] = useState<Product[]>([]);
-
-  useEffect(() => {
-    const fetchProducts = async () => {
-      const response = await fetch("http://127.0.0.1:8000/products");
-      const data = await response.json();
-      const filteredProducts = data.filter(
-        (product: Product) => product.store_id === selectedGroup
-      );
-      setProducts(filteredProducts);
-    };
-
-    fetchProducts();
-  }, [selectedGroup]);
-
   const handleQuantityChange = (product: Product, quantity: number) => {
     onProductSelect({ ...product, quantity }, quantity);
   };
